feat: added review flow, support ticket flow, category flow, featured flow; modified lesson resource, course progress, quiz/instructor deletion

+++ apps/api/src/main/java/com/gii/api/model/response/instructor/InstructorCourseSnapshotResponse.java
@@ -14,13 +14,13 @@ public record InstructorCourseSnapshotResponse(
     PublishStatus status, // DRAFT, PUBLISHED, ARCHIVED
 
     // Enrollment metrics
-    Integer totalEnrolledStudents,
-    Integer completedStudents,
+    Long totalEnrolledStudents,
+    Long completedStudents,
 
     // Content metrics
-    Integer totalSections,
-    Integer totalLessons,
-    Integer liveClassCount,
+    Long totalSections,
+    Long totalLessons,
+    Long liveClassCount,
     Integer quizCount,
 
     // Instructor role
@@ -33,4 +33,4 @@ public record InstructorCourseSnapshotResponse(
     // Quick actions
     String editUrl, // Link to edit course
     String analyticsUrl // Link to course analytics
-) {}
+    ) {}
